requirements.txt
Django==1.10.5
wsgiref

psycopg2
pillow
django-imagekit==3.3
mailchimp
facebook-sdk
django-extensions
django-autoslug
django-summernote
